requirements.txt
erniebot==0.4.0
fastapi==0.104.1
joblib==1.3.2
librosa==0.10.1
numpy==1.26.2
Pillow==10.1.0
scikit_learn==1.3.2
soundfile==0.12.1
torch==2.1.1
torchvision==0.16.1
uvicorn==0.24.0.post1
python-multipart